glibc error?
On Ubuntu, the pingpong test app suffers from lots and lots of jitter,
and the total count of sent messages vary a lot because of this.

On Gentoo, glibc 2.22-r1, there's still some jitter, but a lot less.

Possible problem: The pingpong app is basically a program which loops
as fast as possible, sending one message back and forth between two
threads. Each thread has a message queue, and each thread uses
pthread_cond_wait() to find entries in the queue.

Arrival of new messages are notified using pthread_mutex_signal(not
broadcast) on an unlocked mutex.