Do proper release build.
Build both debug and release builds. This forced us to change the
API a little, so we can return error conditions. The code is
suddenly slightly uglier due to the void** arg and associated casting,
but all in all, we need the error checking.

We also needed to tweak the debug() macro and associated code, to
avoid linking errors when mixing debug and non-debug versions.

PS: More of the same to come.