Created the meta_ssl ADT, which for now is just a copy of meta_sock.
Cloned and renamed the meta_sock ADT, and prefixed all meta_sock
functions with sock_. All meta_ssl functions are prefixed with
ssl_.

The plan? It'd be nice to have some kind of pluggable API for
this thing, so the callers can decide what kind of network they
want, and just instanciate the proper type.