sock: We now have a 'socket factory' in gensocket.c
We can create instances of either straight TCP sockets or SSL sockets,
and then call the same functions for both types.

ATM, the callers still use tcpsocket_, and sslsocket implementation
is just a copy of the tcpsocket implementation. But now we have the
framework in place. Upcoming commits will first start using the
framework, and then implement SSL specific stuff.