ssl: getting closer
Echoserver's callback function gets called, which means that we have
some kind of socket up and running. Reading from it fails, but that's OK
since we dont' handle SSL errors and handshakes properly ATM.

Also, fixed an error in the non-ssl portion of the socket code. There's
no way that could've been an old error. Doesn't matter, it's not an
error anymore.