bitset: Give source some love.
Inline more functions to improve speed and as anti-reversing measure.

Not sure if I want to use gcc function attributes or not. They seem to
be handy, but really uglifies the code. The debug versions should catch
most cases, as long as one has proper tests. The assert() macros will
catch invalid arguments even if the compiler doesn't warn.