Added workaround for conceptual 'bug' in http_response receive.
Problem is that we read n bytes of data from the socket, but we
have no idea if the data is a string or not. Hence we didn't
terminate the buffer and hence it was hard to use the buffer as
a string.

This commit adds a nul character at the end of the buffer read
from the socket, so we can use it as a string. This is kinda
harmless if the content is binary, but we should modify/amend
the ADT's API somehow.