Adding test for helloworld. It fails...
It fails because it runs the executable in one directory,
but the executable expects the SSL certs to be found in
"current directory", as in its source code directory.

I have not found a good solution yet, so imma letting the test fail
until I get a good idea. I want to avoid putting the certs outside
the git-repo in development.