Make socket_t long lived.
Prior to this commit, the accept thread malloced and freed socket_t
instances, and socket_close() freed stuff too. We want to avoid that
to enable seccomp() limits on sbrk, so we move the malloc/free to
the init thread instead. A connection now has a socket_t object for
as long the connection lives.
We use socket_t.fd as a state indicator. -1 means closed. The object
itself is still of course allocated.