seccomp() related.
It's not possible to use seccomp on callback functions called
by the worker threads in the work pool, because that can lead
to one callback limiting other callback functions' needs.

Imagine if handle_robots_txt() restricted itself to IO, but
handle_foo() required malloc and multithreading. So there's
no way around this problem.